Rewrite refs that we have information for in compute_patch_set.
compute_patch_set will now keep track of a map from snapshot IDs to instance IDs
and use it to rewrite Ref properties before returning.

This is only half of proper ref support for Rojo, and won't meaningfully improve
the functionality of `rojo build`, only serving, which doesn't work in
reconciler-mk5 yet.

Since we don't know how to map snapshot IDs to instance IDs for instances that
don't exist yet, we need to defer that step to apply_patch_set, which will be a
follow-up change.