Enable HttpService if running from BuiltInPlugins.
If the rojo plugin is running locally in Roblox Studio's BuiltInPlugins folder, then it has access to enable the HttpService automatically. Roblox added methods for manipulating its value for the Game Settings plugin. This commit adds code to try and enable the HttpService if the module detects that its running from the BuiltInPlugin security context level.