Update README with a clearer description of Rojo
The current introductory description to Rojo focuses more on what the use of Rojo can accomplish rather than what it clearly does. This PR adds a description that is similar to the one in Rojo's documentation by explaining the syncing aspect of it and then explaining how Rojo can be used by Roblox developers to utilize industry-leading software engineering tools.